check for ':>"${BLOCKLIST}"' to occur has been changed from 'test -f' to 'test -e', then 'test -f' occurs after.
this will better handle situations where something is present at
'${BLOCKLIST}' that is not a file--instead of attempting to clobber,
(which will probably fail), an exit with a meaningful error message
happens instead. in the case where nothing was present before and
genblock made a blank file, the script will continue on as normal.