I think it's fine this way.
The code gets more complicated when you try to avoid that defined check
in &wanted, and in any case &wanted is called for every file/directory
anyway. Replacing the check with a subroutine argument and moving the
conditional into the directory check section just means it's checking
for the presence of a parameter instead on each call.