Complete TODO (sanity checking mtu + rtable).
These checks are both shamelessly taken from the OpenBSD source code. A
legal rtable can be between 0 and 255, and sys/net/if_wg.c only accepts
a range between 1-9000 for the MTU.

We want the program to bail before attempting configuration if the
values are bogus, rather than fail an ifconfig(8) and leave networking
in a partial state.

Eventually random_wg would be able to detect that it would leave the
network in a partial state and revert changes. For now, this is better
than nothing.