Change program name. It serves a new purpose.
The difference is that the new program takes WireGuard configuration
files and converts them to hostname.if(5) format. This seems more
broadly useful than doing all of the network configuration work in one
script, as it can be passed on to netstart(8). If OpenBSD does
additional things in netstart, I won't have to play catch-up.

One thing the script didn't do before is install IPv6 routes the way
that netstart does. Using this approach, it's no longer my problem as
that'll happen when the interface is brought up. It also passes a lot of
responsibility for security back to OpenBSD which I trust more than
myself.