Add random_translate_wg
This program selects a random WireGuard file to translate to
hostname.if(5) format, and then delegates that task to
translate_wg_conf(8).

Maybe they can be merged in the future when I learn more about modular
programming.

The man page for random_translate_wg should be added in the next commit.