genblock: performance fix for reduce_domains.
Rather than looping through every single character and checking if the
next one is a period, just split at the periods to begin with. I'm a bit
in disbelief that I didn't do this before.

Here's how impactful this change is.

$ wc -l /etc/blocklist.txt
  663415 /etc/blocklist.txt

$ time ./genblock -t rpz /etc/blocklist.txt > old_genblock_output
    0m21.53s real     0m21.26s user     0m00.30s system

$ time ./genblock -t rpz /etc/blocklist.txt > new_genblock_output
    0m07.74s real     0m07.44s user     0m00.30s system

$ diff -s old_genblock_output new_genblock_output
Files old_genblock_output and new_genblock_output are identical